fix: Fix Rust compilation errors
- Add block_clear() method to Screen struct for clearing rectangular regions
- Fix cursor.style type mismatch (i32, not u8)
- Update do_decset_or_reset() to use correct block_clear() call

+++ terminal-emulator/src/main/rust/src/lib.rs
@@ -195,7 +195,7 @@ pub extern "system" fn Java_com_termux_terminal_TerminalEmulator_getCursorStyleF
 pub extern "system" fn Java_com_termux_terminal_TerminalEmulator_setCursorStyleFromRust(_env: JNIEnv, _class: JClass, ptr: jlong, cursor_style: jint) {
     if ptr == 0 { return; }
     let context = unsafe { &mut *(ptr as *mut TerminalContext) };
-    context.engine.state.cursor.style = cursor_style as u8;
+    context.engine.state.cursor.style = cursor_style as i32;
 }
 
 #[unsafe(no_mangle)]
